Jira 190-191-200
added a post to add favorite teams to db. A get file to have the favorite team always populated when viewing the profile page. Changed the look of the favorite team modal. and removed the activities tab in the profile page. Finally i gave a fixed positiion for the team logo